fix title and legacy heading naming
