+++ src/components/Heading/__tests__/heading-snapshot-tests.jest.js
@@ -1,44 +1,44 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Title from "../Title";
+import Heading from "../Heading";
 
 describe("Text renders correctly", () => {
   it("with empty props", () => {
-    const tree = renderer.create(<Title />).toJSON();
+    const tree = renderer.create(<Heading />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with type", () => {
-    const tree = renderer.create(<Title type="h2">text</Title>).toJSON();
+    const tree = renderer.create(<Heading type="h2">text</Heading>).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with weight", () => {
-    const tree = renderer.create(<Title weight="bold">text</Title>).toJSON();
+    const tree = renderer.create(<Heading weight="bold">text</Heading>).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with color", () => {
-    const tree = renderer.create(<Title color="secondary">text</Title>).toJSON();
+    const tree = renderer.create(<Heading color="secondary">text</Heading>).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with element", () => {
-    const tree = renderer.create(<Title element="p">text</Title>).toJSON();
+    const tree = renderer.create(<Heading element="p">text</Heading>).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with align", () => {
-    const tree = renderer.create(<Title align="center">text</Title>).toJSON();
+    const tree = renderer.create(<Heading align="center">text</Heading>).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it("with element as children", () => {
     const tree = renderer
       .create(
-        <Title>
+        <Heading>
           <div>text</div>
-        </Title>
+        </Heading>
       )
       .toJSON();
     expect(tree).toMatchSnapshot();
