+++ src/components/IconButton/__stories__/IconButton.stories.mdx
@@ -28,7 +28,7 @@ import {
 import { createStoryMetaSettingsDecorator } from "../../../storybook";
 import { Link, createComponentTemplate } from "vibe-storybook-components";
 import Icon from "../../Icon/Icon";
-import Heading from "../../Heading/Heading";
+import Heading from "../../LegacyHeading/LegacyHeading";
 import classes from "./iconButton.stories.module.scss";
 import Avatar from "../../Avatar/Avatar";
