+++ src/components/LegacyHeading/LegacyHeading.module.scss
@@ -1,0 +1,70 @@
+@import "../../styles/themes";
+@import "../../styles/typography";
+@import "../../styles/global-css-settings";
+
+@include create-title-classes;
+@include create-text-classes;
+
+.headingComponent {
+  border: 1px solid transparent;
+  border-radius: $border-radius-small;
+  min-width: 0;
+  @include heading-padding();
+}
+
+.headingComponent.multiLineEllipsis {
+  @include line-clamp(var(--heading-clamp-lines));
+}
+
+/* This is needed to support ellipsis in the middle of the word */
+
+.headingComponent.singleLineEllipsis {
+  @include ellipsis();
+}
+
+.headingComponent.suggestEditOnHover:hover {
+  cursor: text;
+  border: 1px solid;
+  @include theme-prop(border-color, ui-border-color);
+}
+
+/* Override tags styles */
+
+.elementTypeH1 {
+  @include vibe-heading("h1", "normal");
+}
+
+.elementTypeH2.sizeLarge,
+.elementTypeH1.sizeSmall,
+.elementTypeH1.sizeMedium {
+  @include vibe-heading("h2", "normal");
+}
+
+.elementTypeH3.sizeLarge {
+  @include vibe-heading("h2", "light");
+}
+
+.elementTypeH2.sizeSmall,
+.elementTypeH2.sizeMedium,
+.elementTypeH3.sizeSmall,
+.elementTypeH3.sizeMedium,
+.elementTypeH4 {
+  @include vibe-heading("h3", "normal");
+}
+
+.elementTypeH5, .elementTypeP {
+  @include vibe-heading("medium", "normal");
+}
+
+.elementTypeH6 {
+  @include vibe-heading("small", "normal");
+}
+
+.elementTypeSmall {
+  @include font-caption();
+}
+
+.elementTypeLink {
+  text-decoration: none;
+  color: var(--link-color);
+}
