+++ src/components/LegacyHeading/__stories__/Heading.stories.module.scss
@@ -1,0 +1,20 @@
+@import "../../../styles/themes";
+
+.empty-state-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  padding: var(--spacing-large);
+}
+
+.page-header_commands {
+  margin-top: var(--spacing-large);
+  & > * {
+    margin-left: var(--spacing-small);
+  }
+}
+.page-header_search {
+  width: 146px;
+}
