+++ src/components/LegacyHeading/__stories__/assets/index.js
