--- src/components/Title/Title.module.scss
@@ -1,7 +1,0 @@
-@import "../../styles/typography";
-
-@include create-title-classes;
-
-.heading {
-  margin: 0;
-}
