--- src/components/Title/TitleConstants.ts
@@ -1,3 +1,0 @@
-export type TitleType = "h1" | "h2" | "h3";
-
-export type TitleWeight = "bold" | "normal" | "light";
