--- src/components/Title/__stories__/Title.stories.mdx
@@ -1,160 +1,0 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { createStoryMetaSettingsDecorator, Tip, AlphaWarning } from "../../../storybook";
-import Flex from "../../Flex/Flex";
-import { createComponentTemplate } from "vibe-storybook-components";
-import Title from "../Title";
-import { OVERFLOW_TITLE_CONTAINER_ID, ONE_LINE_ELLIPSIS_TEST_ID } from "../__tests__/title-tests-constants";
-import styles from "./Title.stories.module.scss";
-import { titleOverflowSuite } from "../__tests__/title.interactions";
-
-export const metaSettings = createStoryMetaSettingsDecorator({
-  component: Title,
-  enumPropNamesArray: ["types", "weights", "colors", { propName: "align", enumName: "align" }], // List enum props here
-  iconPropNamesArray: [], // List props that are typed as icons here
-  actionPropsArray: [] // List the component's actions here
-});
-
-<Meta title="Text/Title" component={Title} argTypes={metaSettings.argTypes} decorators={metaSettings.decorators} />
-
-<!--- Component template -->
-
-export const textTemplate = createComponentTemplate(Title);
-
-# Title
-
-<AlphaWarning />
-
-<!--- Component documentation -->
-
-- [Overview](#overview)
-- [Props](#props)
-- [Usage](#usage)
-- [Variants](#variants)
-- [Do’s and don’ts](#dos-and-donts)
-- [Use cases and examples](#use-cases-and-examples)
-- [Related components](#related-components)
-- [Feedback](#feedback)
-
-## Overview
-
-Title components are used to title any page sections or sub-sections in top-level page sections.
-
-<Canvas>
-  <Story name="Overview" args={{ children: "Title" }}>
-    {textTemplate.bind({})}
-  </Story>
-</Canvas>
-
-## Props
-
-<ArgsTable story="Overview" />
-
-## Usage
-
-<UsageGuidelines guidelines={["Usage guideline 1", "Usage guideline 2", "Usage guideline 3"]} />
-
-<Tip title="Not what you were looking for?">Tip content</Tip>
-
-## Variants
-
-### Types and weights
-
-<Canvas>
-  <Story name="Types and weights">
-    <Flex
-      gap={Flex.gaps.SMALL}
-      direction={Flex.directions.COLUMN}
-      justify={Flex.justify.START}
-      align={Flex.align.START}
-    >
-      <Title type="h1" weight="bold">
-        Bold H1 title
-      </Title>
-      <Title type="h1" weight="normal">
-        Normal H1 title
-      </Title>
-      <Title type="h1" weight="light">
-        Light H1 title
-      </Title>
-      <Title type="h2" weight="bold">
-        Bold H2 title
-      </Title>
-      <Title type="h2" weight="normal">
-        Normal H2 title
-      </Title>
-      <Title type="h2" weight="light">
-        Light H2 title
-      </Title>
-      <Title type="h3" weight="bold">
-        Bold H3 title
-      </Title>
-      <Title type="h3" weight="normal">
-        Normal H3 title
-      </Title>
-      <Title type="h3" weight="light">
-        Light H3 title
-      </Title>
-    </Flex>
-  </Story>
-</Canvas>
-
-### Colors
-
-<Canvas>
-  <Story name="Colors">
-    <Flex direction={Flex.directions.COLUMN} align={Flex.align.START} gap={Flex.gaps.SMALL}>
-      <Title color="primary">Primary title</Title>
-      <Title color="secondary">Secondary title</Title>
-      <Title element="div" className={styles.primaryBackground} align="center" color="onPrimary">
-        On primary title
-      </Title>
-      <Title element="div" className={styles.invertedBackground} align="center" color="onInverted">
-        On inverted title
-      </Title>
-    </Flex>
-  </Story>
-</Canvas>
-
-### Overflow
-
-Our Text component support overflow state.
-When the text is too long and the component includes an ellipsis flag, we will cut the end of the heading and display instead of it "...".
-
-<Canvas>
-  <Story name="Overflow" play={titleOverflowSuite}>
-    <Flex
-      id={OVERFLOW_TITLE_CONTAINER_ID}
-      direction={Flex.directions.COLUMN}
-      gap={Flex.gaps.MEDIUM}
-      align={Flex.align.STRETCH}
-      style={{ width: "70%" }}
-    >
-      <Title ellipsis={false}>
-        Title without overflow title without overflow title without overflow title without overflow title without
-        overflow title without overflow title without overflow
-      </Title>
-      <Title
-        ellipsis
-        /**for testing purposes**/
-        data-testid={ONE_LINE_ELLIPSIS_TEST_ID}
-        tooltipProps={{
-          containerSelector: `#${OVERFLOW_TITLE_CONTAINER_ID}`
-        }}
-      >
-        Title with overflow title with overflow title with overflow title with overflow title with overflow title with
-        overflow title with overflow title with overflow title with overflow title with overflow
-      </Title>
-      <div>
-        <Title ellipsis withoutTooltip maxLines={2}>
-          Title with overflow title with overflow title with overflow title with overflow title with overflow title with
-          overflow title with overflow title with overflow title with overflow title with overflow title with overflow
-          title with overflow title with overflow title with overflow title with overflow title with overflow
-        </Title>
-      </div>
-    </Flex>
-  </Story>
-</Canvas>
-
-## Related components
-
-<RelatedComponents componentsNames={[]} />
