--- src/components/Title/__stories__/Title.stories.module.scss
@@ -1,15 +1,0 @@
-@import "../../../styles/themes";
-
-@mixin color-rectangle-example {
-  width: 400px;
-  padding: var(--spacing-small);
-}
-.primaryBackground {
-  @include color-rectangle-example;
-  background-color: var(--primary-color);
-}
-
-.invertedBackground {
-  @include color-rectangle-example;
-  background-color: var(--inverted-color-background);
-}
